infra/requirements.txt
attrs==21.2.0
aws-cdk.assets==1.119.0
aws-cdk.aws-apigateway==1.119.0
aws-cdk.aws-applicationautoscaling==1.119.0
aws-cdk.aws-autoscaling-common==1.119.0
aws-cdk.aws-certificatemanager==1.119.0
aws-cdk.aws-cloudformation==1.119.0
aws-cdk.aws-cloudwatch==1.119.0
aws-cdk.aws-codeguruprofiler==1.119.0
aws-cdk.aws-codestarnotifications==1.119.0
aws-cdk.aws-cognito==1.119.0
aws-cdk.aws-dynamodb==1.119.0
aws-cdk.aws-ec2==1.119.0
aws-cdk.aws-ecr==1.119.0
aws-cdk.aws-ecr-assets==1.119.0
aws-cdk.aws-efs==1.119.0
aws-cdk.aws-elasticloadbalancingv2==1.119.0
aws-cdk.aws-events==1.119.0
aws-cdk.aws-iam==1.119.0
aws-cdk.aws-kinesis==1.119.0
aws-cdk.aws-kms==1.119.0
aws-cdk.aws-lambda==1.119.0
aws-cdk.aws-lambda-event-sources==1.119.0
aws-cdk.aws-logs==1.119.0
aws-cdk.aws-route53==1.119.0
aws-cdk.aws-s3==1.119.0
aws-cdk.aws-s3-assets==1.119.0
aws-cdk.aws-s3-notifications==1.119.0
aws-cdk.aws-sam==1.119.0
aws-cdk.aws-secretsmanager==1.119.0
aws-cdk.aws-signer==1.119.0
aws-cdk.aws-sns==1.119.0
aws-cdk.aws-sns-subscriptions==1.119.0
aws-cdk.aws-sqs==1.119.0
aws-cdk.aws-ssm==1.119.0
aws-cdk.cloud-assembly-schema==1.119.0
aws-cdk.core==1.119.0
aws-cdk.custom-resources==1.119.0
aws-cdk.cx-api==1.119.0
aws-cdk.region-info==1.119.0
backports.entry-points-selectable==1.1.0
cattrs==1.7.1
constructs==3.3.132
distlib==0.3.2
filelock==3.0.12
# Editable Git install with no remote (infra==0.0.1)
-e /Users/lan.yao/Repo/etl-pipeline/infra/infra
iniconfig==1.1.1
jsii==1.33.0
packaging==21.0
platformdirs==2.2.0
pluggy==0.13.1
publication==0.0.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.7.2
